# fix: improve performance of sbom/advisory endpoint