--- a/modules/fundamental/src/sbom/service/sbom.rs
+++ b/modules/fundamental/src/sbom/service/sbom.rs
@@ -684,7 +684,11 @@ impl FromQueryResult for QueryCatcher {
             sbom_node: Self::from_query_result_multi_model(res, "", sbom_node::Entity)?,
             context_cpe: Self::from_query_result_multi_model_optional(res, "", cpe::Entity)?,
             status: Self::from_query_result_multi_model(res, "", status::Entity)?,
-            organization: Self::from_query_result_multi_model_optional(res, "", organization::Entity)?,
+            organization: Self::from_query_result_multi_model_optional(
+                res,
+                "",
+                organization::Entity,
+            )?,
         })
     }
 }